+++ cli_tool_audit/call_tools.py
@@ -15,59 +15,50 @@ import subprocess  # nosec
 from dataclasses import dataclass
 from typing import Optional
 
-import toml
-
 # pylint: disable=no-name-in-module
 from whichcraft import which
 
+from cli_tool_audit.config_reader import read_config
 from cli_tool_audit.known_swtiches import KNOWN_SWITCHES
 
 logger = logging.getLogger(__name__)
 
 
-def read_config(file_path: str) -> dict[str, dict[str, str]]:
-    """
-    Read the cli-tools section from a pyproject.toml file.
-
-    Args:
-        file_path (str): The path to the pyproject.toml file.
-
-    Returns:
-        dict[str, dict[str, str]]: A dictionary with the cli-tools section.
-    """
-    # pylint: disable=broad-exception-caught
-    try:
-        with open(file_path, encoding="utf-8") as file:
-            pyproject_data = toml.load(file)
-        return pyproject_data.get("tool", {}).get("cli-tools", {})
-    except BaseException as e:
-        logger.error(e)
-        print(f"Error reading pyproject.toml: {e}")
-        return {}
-
-
 @dataclass
 class ToolAvailabilityResult:
+    """
+    Dataclass for the result of checking tool availability.
+    """
+
     is_available: bool
     is_broken: bool
     version: Optional[str]
 
 
-def check_tool_availability(tool_name: str, version_switch: str = "--version") -> ToolAvailabilityResult:
+def check_tool_availability(
+    tool_name: str, version_switch: str = "--version", only_check_existence: bool = False
+) -> ToolAvailabilityResult:
     """
-    Check if a tool is available in the system's PATH.
+    Check if a tool is available in the system's PATH and if possible, determine a version number.
 
     Args:
         tool_name (str): The name of the tool to check.
         version_switch (str): The switch to get the tool version. Defaults to '--version'.
+        only_check_existence (bool): Only check if the tool exists, don't check version. Defaults to False.
 
     Returns:
         ToolAvailabilityResult: An object containing the availability and version of the tool.
     """
     # Check if the tool is in the system's PATH
     is_broken = True
+
+    # This is wrong?
     if not which(tool_name):
+        logger.warning(f"{tool_name} is not on path.")
         return ToolAvailabilityResult(False, True, None)
+    if only_check_existence:
+        logger.debug(f"{tool_name} exists, but not checking for version.")
+        return ToolAvailabilityResult(True, False, None)
 
     if version_switch is None or version_switch == "--version":
         # override default.
@@ -78,25 +69,31 @@ def check_tool_availability(tool_name: str, version_switch: str = "--version") -
 
     # pylint: disable=broad-exception-caught
     try:
-        result = subprocess.run(
-            [tool_name, version_switch], capture_output=True, text=True, timeout=5, shell=False, check=True
-        )  # nosec
-        out = result.stdout.strip()
-        if "\n" in out:
-            version = out.split("\n")[0]
-        else:
-            version = result.stdout.strip()
+        command = [tool_name, version_switch]
+        result = subprocess.run(command, capture_output=True, text=True, timeout=15, shell=False, check=True)  # nosec
+        # Sometimes version is on line 2 or later.
+        version = result.stdout.strip()
+        if not version:
+            # check stderror
+            logger.debug("Got nothing from stdout, checking stderror")
+            version = result.stderr.strip()
+
+        logger.debug(f"Called tool with {' '.join(command)}, got  {version}")
         is_broken = False
-    except BaseException as exception:
+    except subprocess.CalledProcessError as exception:
         is_broken = True
-        # TODO: run help and parse that?
-        # logger.error(exception)
-        print(exception)
+        logger.error(f"{tool_name} failed invocation with {exception}")
+        logger.error(f"{tool_name} stderr: {exception.stderr}")
+        logger.error(f"{tool_name} stdout: {exception.stdout}")
+    except FileNotFoundError:
+        logger.error(f"{tool_name} is not on path.")
+        return ToolAvailabilityResult(False, True, None)
 
     return ToolAvailabilityResult(True, is_broken, version)
 
 
 if __name__ == "__main__":
+    check_tool_availability("isort")
 
     def run() -> None:
         """Example"""