+++ CHANGES.md
@@ -5,13 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.0.0] - 2024-01-13
+## [1.0.2] - 2024-01-14
 
 ### Added
+- Color mode actually works now, problems are in red.
+- Logging and verbose switch created
 
-- Application created.
+### Fixed
+- When calling subprocess, it now checks stderr if nothing returned by stdout
 
-## [1.0.0] - 2024-01-13
+
+## [1.0.1] - 2024-01-13
 
 ### Added
 - New stress test using npm.
@@ -23,3 +27,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Version switch always defaults to `--version`
 - Tuple results are now dataclasses
 - validate is now a function and exported
+
+## [1.0.0] - 2024-01-13
+
+### Added
+
+- Application created.
+